+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NEXT
 
+- (Feature) Added `SharedSettings#redeemInvitationURL`, which adds the share,
+  secret, and servers from an Earthstar invitation URL.
+- (Feature) Added `createInvitationURL` and `parseInvitationURL` utilities for
+  creating and parsing Earthstar invitation URLs.
 - (Fix) `parseShareAddress` now validates share addresses more strictly so as to
   ensure their suffix is a pubkey.
 - (Fix) Deprecate `generateShareAddress`